+++ pages_plugs/system/menu/form/addMenuByJson.vue
@@ -81,7 +81,8 @@ export default {
 							]
 						},
 						{ key: "menus", title: "JSON内容", type: "textarea",
-							autosize: { minRows: 18, maxRows: 18 }
+							autosize: { minRows: 18, maxRows: 18 },
+							tips:"必须是数组对象形式，JSON结构与static_menu目录下的menu.json文件内容一致。支持直接将静态菜单导入到菜单管理。"
 						},
 					],
 					// 表单验证规则